Allow unlinked notification dispatch
The `dispatch_async` function of Ravenx creates an async task to
dispatch the given notification. This behaviour ties the notification
dispatch process to the caller process, allowing the caller to receive
the notification dispatch status and possible errors.

This may not be desired for some kind of notifications in which a
failure is tollerable. I've added a `dispatch_nolink` function which
does the nofication dispatch in an unlinked task. This allows the
notification to be dispatched in a separated process and fail without
making also fail the caller process.